v0.143 - Multi-Operation Zones Complete
This update completes the refactor of the mission generation algorithm and adds back in the functions that cause the far enemy groups to collapse back towards the center of their zone when a WEST unit enters it and collapse ALL groups to the players when the zone goes under a certain threshold of units. These functions ensure that the players won't have to spend hours looking for one zombie that decided to trek just along the outside of the zone all game, but were disabled during the refactor.

Changelog:
ADDED:
- two new variables are now attached to zone logic objects: _nearEnemyGroups and _farEnemyGroups. These hold groups spawned within the Near Radius and between Near Radius and 750m, respectively
- Triggers to initiate allEnemiesChase.sqf and farEnemiesCollapse.sqf have been recreated for all zones
- 2x POG patches added to description.ext in preparation for assigning to unit at a later date

CHANGED:
- allEnemiesChase.sqf and farEnemiesCollapse.sqf have been significantly modified to allow the new refactored mission generation

TO DO/KNOWN BUGS:
- Additional Multiplayer testing required