v0.25 hotfix b - DynamicSim, Diagnostic Polling and allEnemiesChase fix
Due to latency when a large number of enemies are active, I have enabled Dynamic Simulation on all enemy groups. This will take some fiddling with due to custom AI, but it seems to have improved the lag.
To be able to tell if we're improving, I have added in a diagnostic polling system that polls every 15 seconds to give a server fps count, unit count, and displays how many units are still under dynamic sim. That will help troubleshoot the worst offenders for lag.

Lastly, I've changed the for look within allEnemiesChase and it is no longer throwing piles of errors. I need to do more testing to see if it is still functioning as intended, but this is a good first step since it was a cause of several hangups.

Changlog:
- Added fn_diagnostics to generate a polling diag_log for performance troubleshooting
- Added dynamic simulation for all enemy groups. Some mods are nicer with dynamicSim than others, though...
- Corrected for look within fn_allEnemiesChase to not throw a million errors