v0.25 hotfix c - Improved destroy task performance + green light generation
Changed the destroy object trigger activation from an allObjects check (quite expensive) to a nearObjects one. This seems to have improved the performance of destroy objectives, especially when there are multiple of them in one mission. More performance checks are needed to verify, however.

Also added a green glow to all destroy and retrieval objects that self-deletes when object disappears. Need to do multiplayer testing on this

Changelog:
- Changed destroy task trigger activation from { ({_x inArea thisTrigger} count allMissionObjects '#explosion' > 0) } to { (count (getPos thisTrigger nearObjects ['#explosion', 10]) > 0 } to improve performance and reduce allObjects checks
- Added a #lightpoint generation system within destroy and retrieve missions in fn_createTask to help highlight location of objective. Also added system to delete lights after task is complete.