v0.30 - New Boss Monster Mission
Summary: Added a new mission type which will spawn a 'boss' monster from the EnemySpawnBoss pool generated in the mod loader. Once the boss monster is eliminated, the zone will be completed. The boss is set to a 'level' equal to the number of players divided by 4, and then modified by the difficulty selection in the parameters. This update resolves parts of Issues #6 and #11

Adds:
- New DHO_fnc_spawnBoss function (within Zone/fn_spawnBoss.sqf) generates a 'boss' type monster from the EnemySpawnBoss list and places it in the specified location

Changes:
- Added Case 3 (Destroy Boss) to fn_createTask.sqf to generate boss missions and supporting triggers, tasks, etc.
- Added _missionType selection of 3 to fn_initLocation.sqf to ensure boss type missions are generated

TODO:
- SP and MP testing
- Balance Boss levels
- Description.ext adds DHO_fnc_spawnBoss