v0.31 Add AI Teammates in Host or Dedi for SP
Summary: This adds a check for someone playing DHO by themselves, but over a dedicated server or in player host. If true, a squad of friendly AI soldiers is spawned to support the solo player. This is a highly requested feature and helps resolve #10

Changelog:
Added: Check to init.sqf if only one player is active. If so, it spawns a number of friendly AI soldiers from a hard coded list based on the difficulty (8 AI for easy or normal, 6 for hard, and 4 for Nightmare).

TODO:
- SP & MP Testing
- Consider adding a check to delete soldiers if a second player JIP
- Consider adding AI respawn when player RTBs