v0.32 - AI Unit Loadout Editing
Summary: Added an addAction option to the friendly AI units that opens an arsenal window and allows the players to set a custom loadout on them. This fixes issue #9 . Also some bugfixes on AI unit spawning in player hosted games. ***NOTE: This update requires editing the mission.sqm file to add the new addaction for SP units in the init field. I may make this happen in init.sqf sometime in the future instead***

Changelog:
Added:
- DHO_fnc_setUnitLoadout is a new function that allows editing of an AI unit's loadout. Under the hood, it works by saving the player's current loadout, then copying the targeted unit's loadout to the player, then finally opening the arsenal. Once the player makes changes and closes the arsenal, the player unit's loadout is copied onto the targeted unit, then the player's loadout is restored. This is primarily activated by an addAction on the targeted unit that only works within 500m of the main base area.

Bugfixes:
- AI units spawned in player hosted MP were getting unreliable spawn locations from BIS_fnc_findSafePos. Changed to simply spawn the units within 1m of the MissionCommander

TODO:
- MP testing
- Consider adding a check to delete soldiers if a second player JIP
- Consider adding AI respawn when player RTBs