v0.33a - Modify Loadout Bugfix pt. 1
Summary: After testing, it was revealed that players could attempt to change the loadouts of other players due to how the Modify loadout addAction was added. This is the first step of fixing that issue by spawning a new set of AI units when a player is alone, whether in SP, local hosted, or dedicated, and then adding the modify command. The next step will be to fix the commands on the editor-placed units.

Changelog:
Changed:
- Now deletes all non-player units in SP mode
- Check in init.sqf to spawn friendly AI modified FROM only multiplayer TO any time there is only one player (including SP). This will ensure that newly renamed Modify Unit Loadout addAction is only applied to AI units.

TODO:
- Part 2 will fix the commands on the editor-placed units.
- Consider adding an addAction to player units to allow them to modify loadout anywhere within the HomeBase area.
- MP Test

Known Issues:
- Dynamic sim not working on some monsters (including the boss, which ruined the boss mission)
- Weird heli issues with the default Arma transport system. Need more checking on that