v0.33b - Helicopter Insertion Bugfix Pt 1
Summary: As per issue #12, the insertion helicopter in spawn has developed an issue with not conducting any other waypoints after the initial insertion. After countless hours attempting to fix this bug (which appears to have been caused by a recent Arma 3 update), I have created a work around which simply replaces the crew with new units and a new group each time it lands. This creates more edge bugs to deal with down the road, but for now this essential feature is restored. As issue #12 is also present in the Helicopter Support module, I have replaced that module with the virtual module until I can work with the BI devs to get this functionality working again.

Changelog:
- New function DHO_fn_heliInsert that handles the helicopter insertion as a separate function that can be remoteExec'd
- Modified helicopter insertion script to delete the previous crew and replace with a new crew 30 second after landing at home base. This is to fix the issue outlined in issue #12
- Removed the init lines from each soldier (a leftover accidentally not fixed from v0.33a)

TO DO:
- Clean up this horrendous code
- Remove debug logging
- Soooooo many edge cases to resolve
- MP testing