v0.34 - Dynamic Simulation Bugfix pt1
Summary: As per issue #13, many units are not having their simulation reenabled by dynamicSim when the players get close, leading to "frozen" monsters. This was especially true for any boss type monsters, thus ruining that mission type. To resolve this, I have created a workaround that involves a large trigger manually reenabling all simulation on every unit when a player comes near a zone. This is a bandaid that will need some refining in future passes

Changelog:
- Ensured enableDynamicSimulation is run on all groups AND units spawned
- Modified DHO_fnc_initLocation to create a trigger area that manually disables DynamicSim and manually reenables simulation on all units in the area.

TODO:
- MP Testing
- Clean up code, possibly move into separate function
- Performance (new triggers polling)
- Consider modifying overall spawn code to utilize trigger to spawn all units into the area only when players are around (will help overall performance for servers at the cost of a lag spike when players near zone)